+++ components/script/script_runtime.rs
@@ -703,7 +703,7 @@ pub fn get_reports(cx: *mut RawJSContext, path_seg: String) -> Vec<Report> {
                 reports.push(Report {
                     path,
                     kind,
-                    size: size as usize,
+                    size: size,
                 })
             };
